ServerSentEvents/tests/submissions/MOSQUITO/field-name-id.html

<!DOCTYPE html>
<html>
<head>
		<title>The steps to process the field given a field name and a field value depend on the field name: The field name is "id":</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The steps to process the field given a field name and a field value depend on the field name: The field name is "id":</h1>
	<p id="w">waiting...</p>
	<div id="log"></div>
  <script>
  	setup({timeout:6000});
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0076');
    lei = null;
    t = async_test("The field name is 'id': set the last event ID buffer to the field value.");
 
    s.onopen = function(e){ 
    	/* Connection opened. */
    };
    	
    s.onmessage = function(e){
    	if(lei == null){
    		lei = e.lastEventId;
    		console.log("lei :"+lei);
    	}else{
    		t.step(function(){assert_true(lei != e.lastEventId)});
    		t.done();
    		document.getElementById("w").innerHTML = "";
    	}
    };
    
    //	console.log(s.testevent);
    s.onerror = function(e){
    	/* Connection closed. */
    };
    

  </script>
  </div>
</body>
</html>
